refactor the code, add one new test and do necessary changes
