+++ e2e/imgbuild/regressions.go
@@ -14,7 +14,6 @@ import (
 	"io/ioutil"
 	"log"
 	"os"
-	"os/exec"
 	"path"
 	"path/filepath"
 	"testing"
@@ -542,80 +541,4 @@ From: alpine:latest
 		}),
 		e2e.ExpectExit(0),
 	)
-}
-
-func (c *imgBuildTests) issue476(t *testing.T) {
-	tmpDir, cleanup := e2e.MakeTempDir(t, c.env.TestDir, "issue-476-", "")
-	defer cleanup(t)
-
-	// running under the mode 1(https://apptainer.org/docs/user/main/fakeroot.html)
-	c.env.RunApptainer(
-		t,
-		e2e.WithProfile(e2e.UserProfile),
-		e2e.WithCommand("build"),
-		e2e.WithArgs("--force", "--fakeroot", fmt.Sprintf("%s/openssh-mode1.sif", tmpDir), "testdata/regressions/issue_476.def"),
-		e2e.ExpectExit(0),
-	)
-
-	// running under the mode 2(https://apptainer.org/docs/user/main/fakeroot.html)
-	// temporally remove executable permission from fakeroot
-	e2e.Privileged(func(t *testing.T) {
-		cmd := exec.Command("chmod", "-x", "/usr/bin/fakeroot")
-		err := cmd.Run()
-		if err != nil {
-			t.Fatal(err)
-		}
-	})(t)
-	c.env.RunApptainer(
-		t,
-		e2e.WithProfile(e2e.FakerootModeTwoProfile),
-		e2e.WithCommand("build"),
-		e2e.WithArgs("--force", fmt.Sprintf("%s/openssh-mode2.sif", tmpDir), "testdata/regressions/issue_476.def"),
-		e2e.WithEnv([]string{"_APPTAINER_E2E_FAKEROOT=1"}),
-		e2e.ExpectExit(255), // because chown will fail
-	)
-	// restore permission to fakeroot
-	e2e.Privileged(func(t *testing.T) {
-		cmd := exec.Command("chmod", "+x", "/usr/bin/fakeroot")
-		err := cmd.Run()
-		if err != nil {
-			t.Fatal(err)
-		}
-	})(t)
-
-	// running under the mode 3(https://apptainer.org/docs/user/main/fakeroot.html)
-	c.env.RunApptainer(
-		t,
-		e2e.WithProfile(e2e.FakerootModeThreeProfile),
-		e2e.WithCommand("build"),
-		e2e.WithArgs("--force", fmt.Sprintf("%s/openssh-mode3.sif", tmpDir), "testdata/regressions/issue_476.def"),
-		e2e.WithEnv([]string{"_APPTAINER_E2E_FAKEROOT=1"}),
-		e2e.ExpectExit(0),
-	)
-
-	// running under the mode 4(https://apptainer.org/docs/user/main/fakeroot.html)
-	// disable the userns first
-	e2e.Privileged(func(t *testing.T) {
-		cmd := exec.Command("sysctl", "-w", "kernel.unprivileged_userns_clone=0")
-		err := cmd.Run()
-		if err != nil {
-			t.Fatal(err)
-		}
-	})(t)
-	// // call the apptainer build
-	c.env.RunApptainer(
-		t,
-		e2e.WithProfile(e2e.FakerootModeThreeProfile),
-		e2e.WithCommand("build"),
-		e2e.WithArgs("--force", fmt.Sprintf("%s/openssh-mode4.sif", tmpDir), "testdata/regressions/issue_476.def"),
-		e2e.ExpectExit(0),
-	)
-	// // restore the userns at last
-	e2e.Privileged(func(t *testing.T) {
-		cmd := exec.Command("sysctl", "-w", "kernel.unprivileged_userns_clone=1")
-		err := cmd.Run()
-		if err != nil {
-			t.Fatal(err)
-		}
-	})(t)
 }
